First working version

+++ src/main/java/ru/timokhin/todolist/controller/TodoController.java
@@ -2,11 +2,12 @@ package ru.timokhin.todolist.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import ru.timokhin.todolist.bean.Todo;
-import ru.timokhin.todolist.bean.User;
 import ru.timokhin.todolist.dao.TodoRepository;
 import ru.timokhin.todolist.dao.UserRepository;
+import ru.timokhin.todolist.model.Todo;
+import ru.timokhin.todolist.model.User;
 
+import java.security.Principal;
 import java.util.Collection;
 
 @RestController
@@ -17,33 +18,25 @@ public class TodoController {
     @Autowired
     private TodoRepository todoRepository;
 
-    @RequestMapping(value = "/{userName}/todo", method = RequestMethod.GET)
-    public Collection<Todo> getAllTodos(@PathVariable String userName) {
-        //User user = userRepository.findUserByName(userName).get();
-        return todoRepository.findByUserName(userName);
+    @RequestMapping(value = "/todo", method = RequestMethod.GET)
+    public Collection<Todo> getAllTodos(Principal principal) {
+        String username = principal.getName();
+        return todoRepository.findByUserName(username);
     }
 
-    @RequestMapping(value = "/{userName}/todo/save", method = RequestMethod.POST)
-    public Todo saveTodo(@PathVariable String userName, @RequestBody Todo todo) {
-        User user = userRepository.findUserByName(userName);//.get();
-
+    @RequestMapping(value = "/todo/save", method = RequestMethod.POST)
+    public Todo saveTodo(Principal principal, @RequestBody Todo todo) {
+        User user = userRepository.findUserByName(principal.getName());
         todo.setUser(user);
-
         return todoRepository.save(todo);
-
-        //return todoRepository.save(new Todo(todo.getText(), user));
-
-        /*if (todo.getId() == null) {
-            return todoRepository.save(new Todo(todo.getText(), user));
-        } else {
-            todoRepository.fin
-        }*/
     }
 
-    @RequestMapping(value = "/{userName}/todo/remove", method = RequestMethod.GET)
-    public void removeTodo(@PathVariable String userName, @RequestParam Long id) {
-        User user = userRepository.findUserByName(userName);//.get();
-
-        todoRepository.delete(id);
+    @RequestMapping(value = "/todo/remove", method = RequestMethod.GET)
+    public void removeTodo(Principal principal, @RequestParam Long id) {
+        User user = userRepository.findUserByName(principal.getName());
+        Todo todo = todoRepository.findOne(id);
+        if (todo != null & todo.getUser().equals(user)) {
+            todoRepository.delete(id);
+        }
     }
 }
